+++ rest_framework_json_api/relations.py
@@ -273,7 +273,10 @@ class ResourceRelatedField(HyperlinkedMixin, PrimaryKeyRelatedField):
                 inflection.singularize(field_name),
                 inflection.pluralize(field_name),
             ]
-            includes = getattr(parent, "included_serializers", {})
+            includes = getattr(parent, "included_serializers", None)
+            if includes is None:
+                return None
+
             for field in field_names:
                 if field in includes.keys():
                     return get_resource_type_from_serializer(includes[field])