+++ rest_framework_json_api/renderers.py
@@ -284,7 +284,10 @@ class JSONRenderer(renderers.JSONRenderer):
 
         current_serializer = fields.serializer
         context = current_serializer.context
-        included_serializers = getattr(current_serializer, "included_serializers", {})
+        included_serializers = getattr(current_serializer, "included_serializers", None)
+        if included_serializers is None:
+            return
+
         included_resources = copy.copy(included_resources)
         included_resources = [
             inflection.underscore(value) for value in included_resources
@@ -692,9 +695,11 @@ class BrowsableAPIRenderer(renderers.BrowsableAPIRenderer):
         included_serializers = []
         already_seen.add(serializer)
 
-        for include, included_serializer in getattr(
-            serializer, "included_serializers", {}
-        ).items():
+        included = serializer.included_serializers
+        if included is None:
+            return []
+
+        for include, included_serializer in included.items():
             included_serializers.append(f"{prefix}{include}")
             included_serializers.extend(
                 cls._get_included_serializers(
@@ -715,7 +720,7 @@ class BrowsableAPIRenderer(renderers.BrowsableAPIRenderer):
         except AttributeError:
             return
 
-        if not hasattr(serializer_class, "included_serializers"):
+        if not serializer_class.included_serializers:
             return
 
         template = loader.get_template(self.includes_template)