+++ tests/test_utils.py
@@ -1,5 +1,4 @@
 import pytest
-from django.db import models
 from rest_framework import status
 from rest_framework.generics import GenericAPIView
 from rest_framework.response import Response
@@ -20,13 +19,12 @@ from rest_framework_json_api.utils import (
 )
 from tests.models import (
     BasicModel,
-    DJAModel,
     ForeignKeySource,
     ForeignKeyTarget,
     ManyToManySource,
     ManyToManyTarget,
 )
-from tests.serializers import BasicModelSerializer, ManyToManyTargetSerializer
+from tests.serializers import BasicModelSerializer
 
 
 def test_get_resource_name_no_view():
@@ -342,37 +340,3 @@ def test_get_related_resource_type_from_plain_serializer_class(
     serializer = PlainRelatedResourceTypeSerializer()
     field = serializer.fields["basic_models"]
     assert get_related_resource_type(field) == output
-
-
-class IncludedSerializersModel(DJAModel):
-    self = models.ForeignKey("self", on_delete=models.CASCADE)
-    target = models.ForeignKey(ManyToManyTarget, on_delete=models.CASCADE)
-    other_target = models.ForeignKey(ManyToManyTarget, on_delete=models.CASCADE)
-
-    class Meta:
-        app_label = "tests"
-
-
-class IncludedSerializersSerializer(serializers.ModelSerializer):
-    included_serializers = {
-        "self": "self",
-        "target": ManyToManyTargetSerializer,
-        "other_target": "tests.serializers.ManyToManyTargetSerializer",
-    }
-
-    class Meta:
-        model = IncludedSerializersModel
-        fields = ("self", "other_target", "target")
-
-
-def test_get_included_serializers():
-    included_serializers = getattr(
-        IncludedSerializersSerializer, "included_serializers", {}
-    )
-    expected_included_serializers = {
-        "self": IncludedSerializersSerializer,
-        "target": ManyToManyTargetSerializer,
-        "other_target": ManyToManyTargetSerializer,
-    }
-
-    assert included_serializers == expected_included_serializers